requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2.0
joblib>=1.2.0
matplotlib>=3.5
seaborn>=0.12
nltk>=3.6
wordcloud>=1.8.1
Pillow>=9.4
tqdm>=4.64
